pyproject.toml
[tool.poetry]
name = "icons"
version = "0.1.0"
description = ""
authors = ["Micah Yeager <myeager@innovativeinc.net>"]

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^9.4.0"
PyYAML = "^6.0"
requests = "^2.28.1"

[tool.poetry.dev-dependencies]
black = {version = "*", extras = ["d"]}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
